restructure tests to afford mock reuse
