+++ mod-source-record-storage-server/src/test/java/org/folio/dao/impl/AbstractDaoTest.java
@@ -17,7 +17,7 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 
-public abstract class AbstractDaoTest extends DaoTestMocks {
+public abstract class AbstractDaoTest {
 
   static final String TENANT_ID = "diku";
