+++ mod-source-record-storage-server/src/test/java/org/folio/services/impl/AbstractServiceTest.java
@@ -1,0 +1,17 @@
+package org.folio.services.impl;
+
+import org.junit.Before;
+
+import io.vertx.ext.unit.TestContext;
+
+public abstract class AbstractServiceTest {
+
+  static final String TENANT_ID = "diku";
+
+  @Before
+  public abstract void createService(TestContext context) throws IllegalAccessException;
+
+  @Before
+  public abstract void mockDaos(TestContext context) throws IllegalAccessException;
+
+}
